Allow to customize `HttpTranslator` via `Clients` API. Expose `HttpTranslator` and related (including those just helpful) types by moving them to root of `lib` — so that API end-users could pack/unpack Proto messages more easily.
