+++ client/lib/validate.dart
@@ -27,7 +27,7 @@
 import 'package:optional/optional.dart';
 import 'package:protobuf/protobuf.dart';
 import 'package:spine_client/spine/validate/validation_error.pb.dart';
-import 'package:spine_client/src/known_types.dart';
+import 'package:spine_client/known_types.dart';
 import 'package:sprintf/sprintf.dart';
 
 /// Validates the given message according to the constraints defined in Protobuf.
